Refactor `RemirrorManager` and rename `EditorWrapper`
Fixes #559
