+++ packages/@remirror/dom/src/dom.ts
@@ -2,16 +2,14 @@ import type {
   AnyCombinedUnion,
   BuiltinPreset,
   EditorState,
-  EditorWrapperOutput,
-  EditorWrapperProps,
   PrimitiveSelection,
   RemirrorContentType,
   SchemaFromCombined,
-  UpdateStateParameter,
 } from '@remirror/core';
-import { EditorWrapper, getDocument, isArray, RemirrorManager } from '@remirror/core';
-import { EditorView } from '@remirror/pm/view';
+import { getDocument, isArray, RemirrorManager } from '@remirror/core';
 import { CorePreset, createCoreManager, CreateCoreManagerOptions } from '@remirror/preset-core';
+
+import { DomFramework, DomFrameworkOutput, DomFrameworkProps } from './dom-framework';
 
 /**
  * Create an editor manager. It comes with the `CorePreset` already available.
@@ -23,20 +21,29 @@ export function createDomManager<Combined extends AnyCombinedUnion>(
   return createCoreManager([...combined], options);
 }
 
-export interface DomEditorWrapperProps<Combined extends AnyCombinedUnion>
-  extends EditorWrapperProps<Combined> {
-  /**
-   * Provide a container element. Which the editor will be appended to.
-   */
-  element: Element;
-}
-
 /**
- * Create the prosemirror editor for the dom environment.
+ * Create the `remirror` editor for the DOM environment.
+ *
+ * ```ts
+ * // Get the element to append the editor to.
+ * const element = document.querySelector('#editor');
+ *
+ * // Create the manager.
+ * const manager = createDomManager([], {});
+ *
+ * // Create the dom editor.
+ * const editor = createDomEditor({ element, manager });
+ *
+ * // Focus at the end of the editor.
+ * editor.focus('end');
+ *
+ * // Insert some text.
+ * editor.commands.insertText('Hello Friend!');
+ * ```
  */
 export function createDomEditor<Combined extends AnyCombinedUnion>(
-  props: DomEditorWrapperProps<Combined>,
-): DomEditorWrapperOutput<Combined> {
+  props: DomFrameworkProps<Combined>,
+): DomFrameworkOutput<Combined> {
   const { stringHandler, onError, manager, forceEnvironment, element } = props;
 
   function createStateFromContent(
@@ -45,7 +52,7 @@ export function createDomEditor<Combined extends AnyCombinedUnion>(
   ): EditorState<SchemaFromCombined<Combined>> {
     return manager.createState({
       content,
-      doc: getDocument(forceEnvironment),
+      document: getDocument(forceEnvironment),
       stringHandler,
       selection,
       onError,
@@ -60,70 +67,14 @@ export function createDomEditor<Combined extends AnyCombinedUnion>(
     : ([props.initialContent ?? fallback] as const);
   const initialEditorState = createStateFromContent(initialContent, initialSelection);
 
-  const wrapper = new DomEditorWrapper<Combined>({
+  const framework = new DomFramework<Combined>({
     createStateFromContent,
     getProps: () => props,
     initialEditorState,
     element,
   });
 
-  return wrapper.output;
-}
+  framework.onFirstRender();
 
-interface DomEditorWrapperOutput<Combined extends AnyCombinedUnion>
-  extends EditorWrapperOutput<Combined> {
-  /**
-   * Call this to cleanup the view.
-   */
-  destroy: () => void;
-}
-
-/**
- * The helper class which makes integrating with the DOM easier.
- */
-class DomEditorWrapper<Combined extends AnyCombinedUnion> extends EditorWrapper<
-  Combined,
-  DomEditorWrapperProps<Combined>
-> {
-  /**
-   * Create the prosemirror EditorView.
-   */
-  protected createView(state: EditorState<SchemaFromCombined<Combined>>, element?: Element) {
-    return new EditorView(element, {
-      state,
-      nodeViews: this.manager.store.nodeViews,
-      dispatchTransaction: this.dispatchTransaction,
-      attributes: () => this.getAttributes(),
-      editable: () => {
-        return this.props.editable ?? true;
-      },
-    });
-
-    this.onChange();
-    this.addFocusListeners();
-  }
-
-  /**
-   * Responsible for managing state updates.
-   */
-  protected updateState(parameter: UpdateStateParameter<SchemaFromCombined<Combined>>): void {
-    const { state, tr, triggerChange = true } = parameter;
-
-    // Update the internal prosemirror state. This happens before we update
-    // the component's copy of the state.
-    this.view.updateState(state);
-
-    if (triggerChange) {
-      this.onChange({ state, tr });
-    }
-
-    this.manager.onStateUpdate({ previousState: this.previousState, state });
-  }
-
-  get output() {
-    return {
-      ...this.editorWrapperOutput,
-      destroy: () => this.onDestroy(),
-    };
-  }
+  return framework.domOutput;
 }
