+++ packages/@remirror/styles/all.css
@@ -425,7 +425,7 @@
   background-color: grey;
 }
 
-.remirror-social-editor-wrapper {
+.remirror-social-editor-container {
   position: relative;
   height: 100%;
 }
@@ -478,7 +478,7 @@
   background-color: grey;
 }
 
-.remirror-wysiwyg-editor-wrapper {
+.remirror-wysiwyg-editor-container {
   position: relative;
   height: 100%;
 }
