+++ packages/@remirror/styles/react-social.css
@@ -189,7 +189,7 @@
   background-color: grey;
 }
 
-.remirror-social-editor-wrapper {
+.remirror-social-editor-container {
   position: relative;
   height: 100%;
 }
