+++ packages/@remirror/styles/src/dom.tsx
@@ -446,7 +446,7 @@ export const reactSocialStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-social-editor-wrapper {
+  .remirror-social-editor-container {
     position: relative;
     height: 100%;
   }
@@ -501,7 +501,7 @@ export const reactWysiwygStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-wysiwyg-editor-wrapper {
+  .remirror-wysiwyg-editor-container {
     position: relative;
     height: 100%;
   }
@@ -1154,7 +1154,7 @@ export const allStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-social-editor-wrapper {
+  .remirror-social-editor-container {
     position: relative;
     height: 100%;
   }
@@ -1207,7 +1207,7 @@ export const allStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-wysiwyg-editor-wrapper {
+  .remirror-wysiwyg-editor-container {
     position: relative;
     height: 100%;
   }
