+++ packages/@remirror/styles/src/styled-components.tsx
@@ -692,7 +692,7 @@ export const reactSocialStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-social-editor-wrapper {
+  .remirror-social-editor-container {
     position: relative;
     height: 100%;
   }
@@ -890,7 +890,7 @@ export const ReactSocialStyledComponent = styled.div`
     background-color: grey;
   }
 
-  .remirror-social-editor-wrapper {
+  .remirror-social-editor-container {
     position: relative;
     height: 100%;
   }
@@ -945,7 +945,7 @@ export const reactWysiwygStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-wysiwyg-editor-wrapper {
+  .remirror-wysiwyg-editor-container {
     position: relative;
     height: 100%;
   }
@@ -1000,7 +1000,7 @@ export const ReactWysiwygStyledComponent = styled.div`
     background-color: grey;
   }
 
-  .remirror-wysiwyg-editor-wrapper {
+  .remirror-wysiwyg-editor-container {
     position: relative;
     height: 100%;
   }
@@ -1872,7 +1872,7 @@ export const allStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-social-editor-wrapper {
+  .remirror-social-editor-container {
     position: relative;
     height: 100%;
   }
@@ -1925,7 +1925,7 @@ export const allStyledCss = css`
     background-color: grey;
   }
 
-  .remirror-wysiwyg-editor-wrapper {
+  .remirror-wysiwyg-editor-container {
     position: relative;
     height: 100%;
   }
@@ -2576,7 +2576,7 @@ export const AllStyledComponent = styled.div`
     background-color: grey;
   }
 
-  .remirror-social-editor-wrapper {
+  .remirror-social-editor-container {
     position: relative;
     height: 100%;
   }
@@ -2629,7 +2629,7 @@ export const AllStyledComponent = styled.div`
     background-color: grey;
   }
 
-  .remirror-wysiwyg-editor-wrapper {
+  .remirror-wysiwyg-editor-container {
     position: relative;
     height: 100%;
   }
